feat(domain), refactor(package): 복권 도메인 추가, 기계와 사용자 기능으로 컨트롤러 분리
