+++ src/main/java/lotto/controller/user/LottoDraw.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.controller.user;
 
 import lotto.exception.LottoNumberException;
 import lotto.view.InputView;
@@ -8,15 +8,15 @@ import java.util.List;
 import static camp.nextstep.edu.missionutils.Console.readLine;
 
 public class LottoDraw {
-    private LottoNumberException lottoNumberException;
-    private InputView inputView;
+    private final InputView inputView;
 
     public LottoDraw(InputView inputView) {
-        lottoNumberException = new LottoNumberException();
         this.inputView = inputView;
     }
 
     private boolean isCorrectNumber(List<String> inputs) {
+        LottoNumberException lottoNumberException = new LottoNumberException();
+
         try {
             lottoNumberException.checkSixInputs(inputs);   //6개 입력
             lottoNumberException.checkSixNumbers(inputs);  //6개 숫자
@@ -31,7 +31,6 @@ public class LottoDraw {
 
     public void drawLotto () {
         List<String> inputs;
-        List<Integer> winningNumbers;
 
         while (true) {
             inputView.showWinningNumbersInputForm();
