+++ src/main/java/lotto/domain/WinningNumber.java
@@ -1,0 +1,16 @@
+package lotto.domain;
+
+import java.util.List;
+
+public class WinningNumber {
+    private final Lotto lotto;
+    private int bonusNumber;
+
+    public WinningNumber(Lotto lotto) {
+        this.lotto = lotto;
+    }
+
+    public void setBonusNumber(int bonusNumber) {
+        this.bonusNumber = bonusNumber;
+    }
+}
